stage commit 2

+++ app/repositories/base_repository.py
@@ -16,24 +16,28 @@ class BaseRepository(Generic[ModelType]):
         return self.db.query(self.model).all()
     
     def create(self, obj_in: Dict[str, Any]) -> ModelType:
+        """Create new record"""
         db_obj = self.model(**obj_in)
         self.db.add(db_obj)
         self.db.commit()
         self.db.refresh(db_obj)
         return db_obj
     
-    def update(self, db_obj: ModelType, obj_in: Union[Dict[str, Any], Any]) -> ModelType:
-        obj_data = obj_in if isinstance(obj_in, dict) else obj_in.dict(exclude_unset=True)
-        for key, value in obj_data.items():
-            if hasattr(db_obj, key) and value is not None:
+    def update(self, id: int, obj_in: Dict[str, Any]) -> Optional[ModelType]:
+        """Update record"""
+        db_obj = self.get(id)
+        if db_obj:
+            for key, value in obj_in.items():
                 setattr(db_obj, key, value)
-        self.db.add(db_obj)
-        self.db.commit()
-        self.db.refresh(db_obj)
+            self.db.commit()
+            self.db.refresh(db_obj)
         return db_obj
     
-    def delete(self, id: Any) -> ModelType:
+    def delete(self, id: Any) -> bool:
+        """Delete record"""
         obj = self.get(id)
-        self.db.delete(obj)
-        self.db.commit()
-        return obj
+        if obj:
+            self.db.delete(obj)
+            self.db.commit()
+            return True
+        return False
